scripts/requirements.txt
praw==7.7.1
openai==1.12.0
requests==2.31.0
python-dotenv==1.0.0
